// src/data/model/StarWarsSearchResponse.ts
export type StarWarsSearchResponse = {
    count: number;
    next: string | null;
    previous: string | null;
    results: CharacterResponse[];
};

export type CharacterResponse = {
    name: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: string[];
    species: string[];
    vehicles: string[];
    starships: string[];
    created: string;
    edited: string;
    url: string;
};